cleanup main and move tests to a new file
